battery_v2: fix sysfs overflow, ioctl lock leak and loop/index bounds
Bound the cap_event sscanf into char[16]; release the /dev/batt_misc ioctl lock when kzalloc fails; widen the ADT write loop counter from u8 so srcSize>255 is not truncated; and guard the step float voltage index against step_charging_status<0.

+++ drivers/battery_v2/mfc_charger.c
@@ -1303,7 +1303,7 @@ static int mfc_auth_adt_write(struct mfc_charger_data *charger, u8 *srcData, int
 	u8 wdata = 0;
 	u8 sBuf[144] = {0,};
 	int ret;
-	u8 i;
+	int i;
 
 	pr_info("%s %s : start to write ADT\n", WC_AUTH_MSG, __func__);
